Multi-Regions Support
The deployment of a service requires the creation of a new service
endpoint into Keystone. To create a new service endpoint the
kolla_keystone_service tests if the service is already registered and
updates the URL if so, or create the new endpoint in other cases. In the
actual implementation, the test only looks at the service and the
interface which makes the test behaving wrong in the context of
multi-regions. In multi-regions, we deploy the same service many times,
hence the actual implementation considers that this is the same service
and updates its endpoint rather than creating a new one. This fix adds a
condition to the test that verifies the region to distinguish two
identical services deployed into two different regions.